requirements.txt
Django==2.2
django-datatable-view==2.1.6
django-datatables-view==1.19.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
requests==2.25.1
numpy
pandas
json5==0.9.5
python-decouple
openpyxl
